Create a window
This is a graphical application so we'll need a window to draw into. The
window is also a source of important events for our application, so
we'll end up building our event loop around the window.

The [winit] crate provides an API for window creation and event
management. It wraps several OS-specific APIs.

This change creates a window and an event loop that will run
indefinitely. The function passed to `run` will be called on each
event. In this case we're only handling the event to close the window
which is sent when a user clicks on the close button.

We can set the `control_flow` variable to determine what happens after
the event is processed. This application doesn't do much for now, so we
default to `ControlFlow::Wait`, which means the application doesn't do
anything between events. Later on we'll switch this to
`ControlFlow::Poll` which lets us do other processing (e.g. updating the
main game loop, rendering animations). See [ControlFlow] for more
options.

[winit]: https://github.com/rust-windowing/winit/tree/v0.28.6
[ControlFlow]: https://docs.rs/winit/0.28.6/winit/event_loop/enum.ControlFlow.html