Draw a triangle
OpenGL renders triangles using a [pipeline][0]. There are multiple
stages of the pipeline and some of them are configurable via
shaders. Shaders are programs written in the OpenGL Shading
Language (GLSL) that tell the OpenGL how to transform vertices into
pixels on the screen.

To draw a triangle on the screen, we have to provide shaders for two of
the pipeline stages: the vertex and fragment stage.

The vertex shader is run for every vertex passed through the
pipeline. The output of the vertex shader specifies the position of the
vertex.

The fragment shader is run for every pixel on the rasterized
triangles. It specifies the color the each pixel.

Shaders must be compiled and linked before they can be used. This
happens at runtime after we load the OpenGL library. To keep this
exapmle simpler, we're skipping all error checking.

To start the pipeline, we call glDrawTriangles and specify how many
vertices to use and that they should be combined as triangles. Since
we're only drawing one triangle, we send three vertices for each point
of the triangle.

OpenGL has ways of passing in vertex data using buffers, but in this
case we're hardcoding the vertex data right into the shader. We operate
in the -1.0 to 1.0 range on the X, Y, and Z axes. The `gl_VertexID`
is an incremental index, so we use that to pick the three different
vertices.

For the fragment shader, we're just hard coding the color to be green
for everything.

[0]: https://www.khronos.org/opengl/wiki/Rendering_Pipeline_Overview