Draw a cube
We're building a world out of cubes, so let's try to render a cube. Our
triangle had 3 vertices, and we can make a square from two
triangles. Six squares make a cube, so we'll need:

3 vertices/triangles * 2 triangles/square * 6 squares/cube
  = 36 vertices/cube

Again, we'll hard-code the vertices right into the shader source. The
cube's bottom left corner on the near side will start at the origin (0,
0, 0) and the far side top-right corner will end at (1, 1, 1). It
doesn't matter yet, but we'll assume that the Z values increase further
into the screen.

When defining the vertices for the two triangles on each face of the
cube, we'll move in counter-clockwise order. This is called the winding
order and will be important if we use [face culling] at a later
stage. Choosing clockwise or counter-clockwise doesn't really matter as
long as it's consistent, but counter-clockwise is the default of
glFrontFace so we'll stick with that.

Even though we're rendering all six sides of the cube, we can only see
the front side from the current view.

[0]: https://www.khronos.org/opengl/wiki/Face_Culling