Transform vertex to world coordinates
The default coordinate space in OpenGL is -1.0 to 1.0 for all
axes. Anything outside of this range won't be rendered. This is called
the *clip space*.

However, specifying positions within this cube can be tedious. It would
be preferable if we could use our own coordinate space. For example, if
our world is 8.0 kilometers long on each side, we may want to specify
positions between 0.0 and 8.0 on each axis. We'll call this the *world
space*.

Transforming positions from the world space to clip space is something
we can do in our vertex shader. For example, if we specify a vertex
positioned at (6.0, 4.0, 2.0) in our world, the vertex shader can
transform it to (0.5, 0.0, -0.5) in clip space automatically.

We can do this transform by multiplying our position vectors by a
transformation matrix. To transform positions on each axis from the (0.0,
8.0) range to the (-1.0, 1.0) range, we can use an [orthographic
projection][0].

Converting between coordinate spaces is something we'll use often and
much has been written about it. The [Multiple Coordinates Spaces
chapter][1] in the game math book is a good read, as is this [Learn
OpenGL section][2].

[0]: https://en.wikipedia.org/wiki/Orthographic_projection#Geometry
[1]: https://gamemath.com/book/multiplespaces.html
[2]: https://learnopengl.com/Getting-started/Coordinate-Systems