Handle shader compilation errors
We have the minimum needed to compile and link our shaders, but if
there's an error in the shader syntax the application just won't render
anything. As our shaders become more complicated, it's invaluable to
have details about what went wrong when things aren't working.

This change checks the compile and link status of the shaders, and if
they weren't successful it will fetch more details.

Since we're dealing with a wrapper over a C API, fetching the error
details is a bit more involved than just returning a string from a
function. We first fetch the size of the error message string, allocate
a buffer to read it in, and then convert the C string to a Rust version.