Move cube around via affine transformation
Our coordinates changed from (-1.0, 1.0) to (0.0, 8.0), but our cube is
still sitting at (0.0, 0.0, 0.0) in the bottom-left corner of the
screen. We hard-coded the cube's vertex positions between (0.0, 0.0,
0.0) and (1.0, 1.0, 1.0).

We could update the cube's vertices to the position we want them in, but
we can also move the cube around via another matrix
transformation.

We'll keep our cube vertices positioned at the origin and call this
coordinate system the *model space*. We want to move this cube to
another location in our *world space*.

We can use a 4x4 translation matrix to move a position by a given
amount.

With this approach, we can reuse the same vertices to draw many cubes in
different positions just by varying the transformation matrix. For a
handful of vertices this may not be that important, but for models with
thousands of vertices this could have significant performance savings.