Use perspective projection
We can move our cube around, but we can't see the sides of it. We're
using orthographic projection which doesn't reveal the depth of an
object.

What we want is a [perspective projection][0]. Objects that are further
away should appear closer to the center of the screen. The perspective
projection divides the X and Y components by the depth Z. Objects
further away (i.e. a larger Z) will have smaller X and Y values and
appear closer to the origin.

Note that we don't do the division explicitly. It happens by setting the
divisor as the fourth component of the vec4 for `gl_Position`. This is
such a common operation in 3D applications that OpenGL handles the
division for us as part of the rendering pipeline.

So a value of `vec4(x, y, z, w)` will yield a position of
`vec3(x / w, y / w, z / w)`.

[0]: https://gamemath.com/book/matrixmore.html#perspective_projection