Change color of each side of cube
Our fragment shader has been assigning the same color to every
pixel. Without any form of lighting, this makes it difficult to
differentiate the different sides of the cube.

Right now the fragment shader doesn't have a way to determine which
pixel belongs to which side, but we can pass it that information from
the vertex shader. We can specify output variables from one
stage and link them to input variables in another.

We'll define six colors in our vertex shader and use the same color for
every six vertices (which define one side of the cube). When OpenGL
transforms these vertices into pixels, it will pass along the associated
color value to the fragment shader.

Since each vertex for a single side has the same color, every pixel for
that side will receive the same color. But we could also assign each
vertex a different color and OpenGL will mix the values appropriately to
build a color gradient.