Enable depth testing
We have the shape of the cube in perspective, but we're seeing the wrong
faces. Faces that should be hidden are rendered on top of faces that
should be visible.

This is partially because when we perform the perspective divide, we're
dividing all components of the position by their depth. Our coordinates
end up being (x / z, y / z, z / z). All of our z coordinates are set to
1.0.

To fix this, we need to map the range of Z values we want to display
onto the -1.0 to 1.0 range in the clipping plane. We can do this by
adjusting our perspective transform. There is a detailed derivation for
how this matrix was produced on [OGLdev][0].

We also need to enable [depth testing][1] in OpenGL. Without depth
testing, OpenGL will display whatever was rendered last in front of any
other rendered object. Depth testing will store the Z value for each
rendered fragment in the depth buffer. When another fragment would be
rendered in the same spot, it will first check the depth buffer to see
if anything is in front of it to decide whether it should be visible.

By default OpenGL will keep depth values that are _less_ than the
currently stored value, although it is possible to change this with
[glDepthFunc][2]. We'll keep this default and assume that the Z axis is
pointing into the screen, so that smaller values are closer than larger
values. This is referred to as a left-handed coordinate space. We could
change it to a right-handed space by either modifying the depth function
or the perspective transform.

[0]: https://ogldev.org/www/tutorial12/tutorial12.html
[1]: https://www.khronos.org/opengl/wiki/Depth_Test
[2]: https://registry.khronos.org/OpenGL-Refpages/gl4/html/glDepthFunc.xhtml
[3]: https://gamemath.com/book/cartesianspace.html#3d_hands