Render multiple cubes
In our game we're going to render many cubes to make up a world. Most of
what we've built can be reused, but we'll want to put the cubes in
different positions.

We've already built a transformation that can move the model cube at the
origin to an arbitrary position in the world. The position is supplied
to the vertex shader via a uniform which we had given a default
value.

The more common use for a uniform is for the application to pass in
values to the shader each time the pipeline is invoked (i.e. when we
call glDrawTriangles). In this case we want to specify a new position
each time we draw a cube.