Add texture to cube
Specifying color by vertex can be limiting. If we want to add more
detail, we'd need to insert additional vertices.

Instead, we can apply a texture to our cube. The texture we'll use a
2-dimensional array the stores color information.

In this case we're creating a small texture with just four pixels. The
[glTexImage2D function][0] lets us specify how we're uploading our data,
and in this case we're giving it a 2x2 matrix of RGB values.

We also need to specify how the texture maps to our cube. In this case,
we're mapping the corners of the texture to the corners of the cube
face.

[0]: https://registry.khronos.org/OpenGL-Refpages/gl4/html/glTexImage2D.xhtml