Textures have sharper edges
The texture we provided only has 4 pixels (2x2). The rendered cubes will
likely have more than 4 pixels per face, so OpenGL will interpolate
between values to create a color gradient.

This isn't what we want for our game aesthetic, so we can instruct
OpenGL to not interpolate but choose the nearest color to the
point. This will result in a sharper, blockier texture which is what
we're looking for.