Add a world representation
Now that we can render some blocks in different locations, lets try to
build a small world.

This change adds a new World struct that provides an iterator for the
blocks in the world. Right now it just produces a flat world that's 20 blocks
width in the X direction and 20 blocks deep in the Z direction.

The X and Z axes start at position 0 and extend in the positive
direction. By default our view of the world is from the origin, so it
should look like we're at the corner. We also set the Y value of the
single layer of blocks to -2 so it appears that we're standing on top of
the world.