Fixed timestep
When we switched to continuous movement, we updated the camera position
every time the event queue was clear. However, we didn't add any
checks to see how often this update occurred.

On some systems, OpenGL will wait when swapping buffers to match up with
the monitor refresh rate. In this case, we're probably updating the
camera 60, 75, 120, or 144 times per second.

If OpenGL does not wait for the refresh, it will run the loop as fast as
possible, which could be many thousands of times per second for such a
simple scene.

The number of times we update our position affects how fast we move. If
my velocity is set to 1.0 and I'm adding that to my position 120 times a
second, I will move twice as fast as if my refresh rate was 60 times per
second. And if my frame rate isn't locked to my refresh rate and the
position is updated thousands of times per second, I'll zoom right past
the blocks.

One solution is to lock down how often we update the world (and the
camera position). Regardless of how many times MainEventsCleared is
called, we only update the world if enough time has elapsed. We can
specify our desired frames per second, and from there determine the
amount of time per frame.

With this change, regardless of the refresh rate or vsync, the camera
will move at the same speed.

See the [Fix Your Timestep][0] post for more options.

[0]: https://gafferongames.com/post/fix_your_timestep/