Print frames per second
In our last change we locked the number of world updates to a fixed
rate. However, we still render everytime we receive MainEventsCleared,
even if the world hasn't changed.

Depending on whether you have vsync enabled, you'll be rendering at the
refresh rate of your monitor, or many thousands of times per second if
not.

We can monitor this by incrementing a counter every frame and printing
out the aggregate periodically. We can use this value as a crude measure
of performance as we try to scale up our world shortly.