Cache uniform locations
In the last commit we profiled our application and generated a
flamegraph. Ignoring the `iridium:gdrv0` thread, we can see where we're
spending our time on the right side of the graph. There's a lot of setup
on the bottom half, until we reach `iridium::main` which is the main
function in our program.

Moving up the stack, we're spending almost all of our time in
`Renderer::draw_cube`. Within that method, a large chunk is consumed
getting the uniform locations within our shaders.

These uniform locations don't change once the program is linked, so we
could try caching them. This change introduces a HashMap to store the
locations by name after their first lookup.

The names are changed to static strings since we're hardcoding them in
the source for now.

Running this on my machine improved performance from ~10 FPS to ~16
FPS for a 400x400 world. That's significant, but we still need to do
quite a bit more. I've included an updated flamegraph from the latest
run.