Cache block positions at startup
In our last change, we were uploading all block positions at once into a
vertex buffer. This drastically cut down on the number of glDraw calls
we needed to make.

However, we're spending a lot of time compiling this list of positions
each frame, even though they aren't changing. The flame graph shows that
we're spending most of our CPU time collecting these into a Rust vector.

With this change, we upload the block positions once at startup and
cache the number of blocks to render. Each frame we just set the camera
position and instruct OpenGL to render the same number of blocks. I'm
able to comfortably render 1000x1000 = 1 million blocks per frame at 60
FPS.

The new flame graph is interesting as well. A large portion is dedicated
to setting up the window and OpenGL config (which does not take a
noticeable amount of time), and the rest is a smattering of OpenGL calls
in the event loop. I suspect this means that our application isn't
spending much time on the CPU, probably because we've offloaded all of
the work to the GPU. Each time we're calling `present`, we're waiting
for the monitor refresh rate to do anything (if VSync is enabled).