Add directional lighting
When we modified the height of our world, it was difficult to see what
changed. Everything has the same texture and blended together into a
green soup.

We can add some basic lighting to help differentiate the sides of the
cubes.

We'll assume the light is coming from the sun, where it shines down at
the same angle for all objects. Any faces on the cube that are facing
the sun will be lit brighter than those that are not facing the sun.

These calculations happen in the fragment shader. We need to know the
angle of the face, which is passed from the vertex shader. This is the
surface normal.

Taking the dot product between the vector facing the sun and the surface
normal will tell us how much the sun is shining on the surface. We'll
discard negative values which mean the surface is facing away from the
sun.

We're ignoring any cells that may obstruct the sun.