Add skybox
I'm tired of looking at our mauve background. We can add some texture to
our sky with a skybox.

The way we draw a skybox is different than our cubes. Rather than
modifying the cube shader, we'll introduce a second skybox shader to
render this.

Using multiple shaders means we need to do more bookkeeping and better
management of binding our objects. OpenGL is an opaque state
machine. Even though the gl functions don't modify a given struct, they
are setting some global state that will persist between frames.

The skybox shader will render a box that stretches from (-1, -1, -1)
to (1, 1, 1), i.e. the entire OpenGL clipping box. The camera will
always be located at the center of the box (i.e. the origin). The camera
can rotate within the box, but as it moves the box moves with it. In
this way the camera can never close to the edges of the box, giving the
appearance of something rendered in the distance.

We'll texture six sides of the box. We pass the viewing angle to the
fragment shader which uses a special function to convert the angle into
a texture coordinate.