Require active program to update uniform
One challenge with multiple programs is that a program needs to be
active before you can update the uniform. If we tried to update a
uniform for all programs at once, only the active one will actually be
updated.

```
self.cube_program.activate();
self.cube_program.set_uniform_f32("camera_pitch", 0.0);

// This won't actually update since the skybox program isn't active
self.skybox_program.set_uniform_f32("camera_pitch", 0.0);
```

We've been lax with our error handling in OpenGL which may have caught
this, but it would also be great if the type system could prevent this.

This change introduces a new `ActiveProgram` struct that's basically a
wrapper for our existing `Program`. The `set_uniform_*` methods are
moved to this new struct, and we can mutably borrow the active program
if we need to update a uniform.

However, if we mutably borrow one program, we can't borrow the other at
the same time (I think). This prevents activating both programs at once.