Heightmap with coarser grid
Previously, when we had chosen a random height for each X-Z coordinate,
the terrain fluctuated wildly. Adjacent blocks had no influence on each
other so the terrain was very jagged.

Instead of choosing a random height for *each* X-Z coordinate, we can
use fewer random values that are further apart, and interpolate the
height of coordinates between these values.

In this example, we're choosing a random height every 16 blocks in the X
and Z directions. This creates a grid pattern with cells that are 16x16
and a random height at the corner of each cell.

Coordinates that are not at the corners will derive their height from
the values at the four corners of the cell.