Press Escape to quit
Before we start rendering larger scenes, we'll switch to a fullscreen
window. But before we go fullscreen, we need an easy way to exit the
program.

This change shuts down the application if the Escape key is
pressed.

Keyboard input events are tied to the active window. In our event loop
we're checking that the key was pressed down and which keycode was
sent.

We'll be expanding the list of keys we check shortly, so we're ignoring
clippy's suggestion to switch the `match` statement to a single `if`.