Generate random grass texture
Instead of hard coding a 2x2 array, generate an array of random bytes
for an 8x8 texture. We're still using gl::GREEN when uploading our
texture data, so the random values might look like a poor quality grass
texture (if you squint hard enough).

For random values there's the rand crate, but I was curious how random
numbers work so I implemented my own. This implementation is a [xorshift
random number generator][0] which is very simple to implement.

When we add more randomness later, we'll likely switch to the rand crate
which is far more robust.

[0]: https://en.wikipedia.org/wiki/Xorshift