Randomize height
Our solid cube is very boring. To add some variation, we can adjust the
height of each coordinate.

This change specifies a max height for each XZ coordinate in our
terrain. The values are randomly chosen between 1 and the y height.

The effect looks like a bunch of noise, but it's a start.