Move continously
The current camera controls move by a fixed amount each time you press a
key. This lets us change the view, but is far from a smooth experience.

This change will continue moving the camera while a key is
pressed. Rather than updating the camera position directly on each key
event, it sets the camera velocity. The velocity will update the
position every frame as long as the user is holding down the key. We now
listen for key release events to know when to reset the velocity.

One important change is the control flow for the event loop. Previously,
we would only do react if an event came in (e.g. a key press). Now we
need to update the camera position every frame, even if the keyboard
state hasn't changed.

Changing the control flow to polling means if there are no events, we'll
receive [MainEventsCleared][0]. Here we can run `world.update` to update
the camera position and redraw the frame.

Note that the [RedrawRequested][1] event is still used when the window
is resized, so we'll keep that handler and adjust the viewport and
aspect ratio when it happens.

Also, since we're continously redrawing we no longer need to explicitly
request a redraw action.

[0]: https://docs.rs/winit/0.28.6/winit/event/enum.Event.html#variant.MainEventsCleared
[1]: https://docs.rs/winit/0.28.6/winit/event/enum.Event.html#variant.RedrawRequested