Move in direction of heading
Even though we can look around with the mouse, when we move forward
we're still traveling along the +Z axis.

Our camera heading is only used in our shader to transform our view of
the world. We also need to transform our velocity before we update our
position each tick.

It's possible to rotate a 3D vector by multiplying it with a [rotational
matrix][0]. This change takes a shortcut by directly applying the matrix
multiplication to the elements of the vector and returning a new vector
that's been rotated.

This is also a good time to start adding unit tests to ensure we're
getting the results we expect. These tests can be run with `cargo test`
and add some initial converage.

We have to define a custom vec3 assertion because there are often small
rounding errors after performing multiple operations on floating point
values. The assert method will ignore these differences when checking if
two vectors are equal.

[0]: https://gamemath.com/book/matrixtransforms.html#rotation_3d_cardinal_axis