Stretch cube to match aspect ratio
While our cubes have unit length on each axis, when they're rendered
they may appear more rectangular.

OpenGL's clipping space ranges from -1.0 to 1.0 on the X and Y axis, but
it has to then map that to screen space coordinates of the window. If
the window isn't a square, it has to stretch the clipping space to
fit. Most windows/monitors are wider than tall, so the cubes will often
be stretched along the X axis.

To compensate for this, we can stretch everything along either the X or
Y axis when we transform from the world space to the clipping
space. Then when OpenGL transforms from clipping space to screen space,
it squeezes it back to the desired shape.

The amount that we need to stretch depends on the dimensions of the
window, specifically how much wider it is than tall. This is referred to
the aspect ratio. We can pass in the aspect ratio as a uniform value
which will then be added to the transform matrix.

It's possible that this value can change if the window is resized, so we
can set this value whenever the window requests a redraw.