Move camera around
As we start building out our world, we have limited ability to explore
it. Our view is stuck at the origin, facing down the +Z axis.

To look around, we can introduce a camera that we can control. The
camera is represented as a point in space (i.e. a 3D vector). To move it
around, we listen for keyboard events and then increment the camera
position in any of the six cardinal directions (up, down, left, right,
forward, backward).

To handle rendering from the viewport of the camera, we introduce the
[camera coordinate space][0]. This space puts the camera at the origin
facing down the +Z axis which is how we've been rendering everything so
far.

If the camera moves away from the origin, we need a way to transform the
world in a way that puts the camera back to (0, 0, 0). This is the
world-to-camera transform that will shift everything in the world
relative to the camera's position. See this section on [coordinate space
transformations][1] for more details.

[0]: https://gamemath.com/book/multiplespaces.html#camera
[1]: https://gamemath.com/book/multiplespaces.html#transformations