Rotate camera left and right
We can move the camera around, but we're always looking forward down the
+Z axis. We want to be able to rotate the camera so we can look at
different angles.

With this change we introduce the camera heading. We're going to use
heading as defined [here]:

* The heading defines the horizontal angle (i.e. looking left or right).
* A heading of 0.0 means facing forward down the +Z axis.
* Increasing the heading will rotate clockwise.

The heading value is in radians, so incrementing the heading by 2 * PI =
6.28 will return the camera to the original value.

To use the heading, we fold it into our world to camera transformation
matrix. First we move the world so that the camera is at the origin,
then we rotate the world so that our camera is facing forward
again. Since we're looking left and right, we're effectively [spinning
the world around the Y axis][1].

To change the heading from user input, we're now listening for [mouse
movement events][2]. These events include the amount the mouse moved in
the X and Y directions, and for now we only care about the X value.

The units of the values are unspecified (maybe pixels?), but adding the
X-delta to the heading value causes the camera to spin way too fast. We
can add in a factor that represents the mouse sensitivity to bring it
down to a reasonable rotation speed.

[0]: https://gamemath.com/book/polarspace.html#polar_conventions
[1]:
https://gamemath.com/book/matrixtransforms.html#rotation_3d_cardinal_axis
[2]:
https://docs.rs/winit/0.28.6/winit/event/enum.DeviceEvent.html#variant.MouseMotion