Look up and down
Our camera heading lets us look around horizontally (i.e. left and
right). To look around vertically, we'll need to capture another angle.

This change introduces the camera pitch which captures the vertical
angle (i.e. up and down). The functionality is similar to the camera
heading, but we're using a different [rotational matrix][0] to spin the
camera around the X axis. We're also using the [conventions outlined
here][1] to define pitch:

* A pitch of 0 is facing forward.
* As pitch increases, the angle faces down.

Note that we're not including pitch in our velocity. Even if you're looking
down or up, you'll still move parallel to the X-Z plane. This is how
we'll likely implement it when we support walking on the world, so
leaving it as-is for now.

Note that we are clamping the pitch between straight up (PI/2) and
straight down (-PI/2). If you exceed those values you would flip the
world upside down, which can be very disorienting.

[0]:
https://gamemath.com/book/matrixtransforms.html#rotation_3d_cardinal_axis
[1]: https://gamemath.com/book/polarspace.html#polar_conventions