Use instancing to send block positions in bulk
Looking at the flame graph again, we're now spending most of our time in
the `draw_cube` method, which calls `set_uniform_vec3` to set the
position uniform, and then `gl::DrawArrays`. Neither of these commands
are doing much, but we're calling them a lot. With a 400x400x1 world,
we're rendering 160,000 cubes per frame. At 60 frames per second, that's
almost 10 million cubes per second.

I don't think modern GPUs would struggle to render 160,000 cubes per
frame. However, there is overhead in sending OpenGL lots of small chunks
of work (as opposed to batching it together). This [StackOverflow
post][0] and the [referenced presentation][1] suggest that the CPU
cannot send data fast enough to the GPU to process. One solution is to
send larger batches of information at a time.

For our world, we're rendering the same cube many times, just in
different positions. We can leverage [instancing][2] in OpenGL to
specify how many cubes we want to draw and supply the positions all
together via a buffer. [Learn OpenGL][3] has an excellent tutorial for
how to do this.

This change involves sending over all the block positions to the
renderer, which will then upload them to a vertex buffer and send to the
GPU via [glBufferData][4]. In the shader, rather than receiving the
position as a unifom, it's coming in as an [input variable][5]. To link
the buffer data to the input variable in the shader, we can use
[glVertexAttribPointer][6] to specify the type and layout of data (3
floats are represented as a vec3 in the shader).

When I run this in release mode (`cargo run --release`), I'm back to a
consistent 60 FPS on my machine. Looking at the updated flamegraph,
we're now spending almost all our time collecting the positions into a
Rust vector each frame. We'll tackle this issue next.

Another interesting note is that the `iridium:gdrv0` thread seems to
have disappeared. The work in that thread likely corresponded to the
overhead of the many `drawArrays` calls that we ended up replacing.

[0]:
https://stackoverflow.com/questions/4853856/why-are-draw-calls-expensive
[1]: https://www.nvidia.com/docs/IO/8228/BatchBatchBatch.pdf
[2]: https://www.khronos.org/opengl/wiki/Vertex_Rendering#Instancing
[3]: https://learnopengl.com/Advanced-OpenGL/Instancing
[4]: https://registry.khronos.org/OpenGL-Refpages/gl4/html/glBufferData.xhtml
[5]: https://www.khronos.org/opengl/wiki/Vertex_Shader#Inputs
[6]: https://registry.khronos.org/OpenGL-Refpages/gl4/html/glVertexAttribPointer.xhtml