Create OpenGL context
TODO update for glutin 0.31.0

To use OpenGL, we need an [OpenGL context][0]. The context is what
stores the current OpenGL state which we manipulate via
glFunctions (e.g. storing the background color set via glClearColor).

There are several platform-specific APIs for creating an OpenGL context:
[WGL][2] for Windows, [CGL][3] for Macs, [GLX][4] for X Windows. There's
also [EGL][5] which is used by Wayland and mobile clients to create
OpenGL ES context.

Fortunately, the [glutin][6] library handles cross-platform context
creation and integrates with winit. There are several steps to the
process.

The four main concepts we'll encounter are:

* the _display_, which roughly corresponds to some connection to a
display server,
* the _surface_, which is something we can draw onto. In our case we are
using a window surface, but there can also be other surfaces for
rendering offscreen.
* the _config_, which describes the surface,
* and the _context_, which stores the current state of OpenGL.

We also need to [load OpenGL functions][7] which we use the [gl
crate][8] for.

We then make our first OpenGL call with [gl::ClearColor][9]. This stores
the color in the active context, so that when we call [gl::Clear][10]
later it will set the screen to a nice amethyst purple.

[0]: https://www.khronos.org/opengl/wiki/OpenGL_Context
[1]: https://registry.khronos.org/EGL/specs/eglspec.1.5.pdf
[2]: https://en.wikipedia.org/wiki/WGL_(API)
[3]: https://en.wikipedia.org/wiki/Core_OpenGL
[4]: https://en.wikipedia.org/wiki/GLX
[5]: https://registry.khronos.org/EGL/specs/eglspec.1.5.pdf
[6]: https://github.com/rust-windowing/glutin
[7]: https://www.khronos.org/opengl/wiki/Load_OpenGL_Functions
[8]: https://crates.io/crates/gl
[9]:
https://registry.khronos.org/OpenGL-Refpages/gl4/html/glClearColor.xhtml
[10]: https://registry.khronos.org/OpenGL-Refpages/gl4/html/glClear.xhtml