Extract render module
There are a lot of details that go into setting up an OpenGL context,
including several unsafe calls. This change extracts everything related
to drawing to a new module to keep the main function simpler.

Right now it doesn't handle any errors, we can improve on that later if
needed.