Fix collision detection along cardinal axes
When converting a position to a global index, the conversion was
truncating the fractional part to determine the index (e.g. 3.14 would
be converted to 3).

This conversion failed when crossing into negative values. We want every
global index to be unit length on each side, but indices at the origin
were double the size since the non-inclusive range (-1.0, 1.0) would
translate to index 0 (i.e. both -0.9999 and 0.9999 would map to index
0).

This was updated to always round down before truncating the fraction, so
-0.9 would be rounded down to -1.0 and converted to index -1.