Left-clicking destroys block under camera
We want to allow destroying blocks. However, our terrain is modeled as a
heightmap, with just a single value for each X-Z location. This means we
can't destroy blocks beneath the top block because we don't have a way
to represent holes like that.

We'll address this issue later, but for now when a user left-clicks the
mouse we'll destroy the top block that's below the camera
position (i.e. decrement the heightmap by 1 at that location).

Since we cache all of the block positions on the GPU, we'll need to
rebuild the cache on any frame that the block is destroyed.